% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{aGreatFunc}
\alias{aGreatFunc}
\title{Add together two numbers.}
\usage{
aGreatFunc(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This is a description
}
\examples{
aGreatfunction(1, 1)
aGreatfunction(10, 1)
}
