% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFunctions.R
\name{aGreatFunction}
\alias{aGreatFunction}
\title{Add together two numbers.}
\usage{
aGreatFunction(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This is a description
}
\examples{
aGreatFunction(1, 1)
aGreatFunction(10, 1)
}
